requirements.txt
asyncpg==0.20.1
discord.py==1.4.1
halo==0.0.29
humanize==2.4.0
psutil==5.7.0
python-dotenv==0.13.0
git+https://github.com/Rapptz/discord-ext-menus
discord~=1.0.1
aiofiles~=0.5.0